requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.0
Django==3.2.4
docopt==0.6.2
future==0.18.2
html5lib==1.1
num2words==0.5.10
Pillow==8.2.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
qrcode==7.1
reportlab==3.5.68
six==1.16.0
sqlparse==0.4.1
webencodings==0.5.1
xhtml2pdf==0.2.5
